ic-flask/requirements.txt
Flask==1.1.2
torch==1.6.0
torchvision==0.7.0
pillow==8.0.0
prometheus-flask-exporter==0.18.1